Fixed issue of ghost illegal move again
Previously, calling checkdeath before and after the ghost move is the same as the pacman calling checkdeath himself. This causes some issues. 

For this version, I set it such that the ghost can only perform its action if it's not eaten. At the next iteration, when pacman makes a move, the isEaten state of the ghost becomes false again. Thus able to move. 

Hopefully this will work.